Move background things to jarabe.model.desktop
